Fixed CR comment
