+++ pkg/network/http/testutil/pythonserver.go
@@ -1,0 +1,126 @@
+package testutil
+
+import (
+	"context"
+	"crypto/tls"
+	"fmt"
+	"github.com/stretchr/testify/require"
+	"io/ioutil"
+	"net"
+	nethttp "net/http"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strconv"
+	"sync"
+	"testing"
+	"time"
+)
+
+const (
+	pythonSSLServerFormat = `import http.server, ssl, sys
+
+YES = ('true', '1', 't', 'y', 'yes')
+
+class RequestHandler(http.server.BaseHTTPRequestHandler):
+    protocol_version = 'HTTP/1.1'
+    daemon_threads = True
+
+    def do_GET(self):
+        status_code = int(self.path.split("/")[1])
+        self.send_response(status_code)
+        self.send_header('Connection', '%s')
+        self.send_header('Content-type', 'application/octet-stream')
+        self.send_header('Content-Length', '0')
+        self.end_headers()
+
+server_address = ('%s', %s)
+httpd = http.server.HTTPServer(server_address, RequestHandler)
+
+if len(sys.argv) >= 2 and sys.argv[1].lower() in YES:
+    httpd.socket = ssl.wrap_socket(httpd.socket,
+                                   server_side=True,
+                                   certfile='%s',
+                                   keyfile='%s',
+                                   ssl_version=ssl.PROTOCOL_TLS)
+try:
+    httpd.serve_forever()
+finally:
+    httpd.shutdown()
+`
+)
+
+var (
+	disableTLSVerification = sync.Once{}
+)
+
+func HTTPPythonServer(t *testing.T, addr string, options Options) (func(), error) {
+	host, port, err := net.SplitHostPort(addr)
+	if err != nil {
+		return nil, err
+	}
+	if options.EnableTLS {
+		disableTLSVerification.Do(func() {
+			nethttp.DefaultTransport.(*nethttp.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
+		})
+	}
+
+	connectionHeaderValue := "close"
+	if options.EnableKeepAlives {
+		connectionHeaderValue = "keep-alive"
+	}
+	curDir, _ := CurDir()
+	crtPath := filepath.Join(curDir, "testdata/cert.pem.0")
+	keyPath := filepath.Join(curDir, "testdata/server.key")
+	pythonSSLServer := fmt.Sprintf(pythonSSLServerFormat, connectionHeaderValue, host, port, crtPath, keyPath)
+	scriptFile, err := writeTempFile("python_openssl_script", pythonSSLServer)
+	require.NoError(t, err)
+	defer scriptFile.Close()
+
+	cmd := exec.Command("python3", scriptFile.Name(), strconv.FormatBool(options.EnableTLS))
+	go require.NoError(t, cmd.Start())
+
+	// Waiting for the server to be ready
+	portCtx, cancelPortCtx := context.WithDeadline(context.Background(), time.Now().Add(time.Second*5))
+	rawConnect(portCtx, t, host, port)
+	cancelPortCtx()
+
+	return func() {
+		if cmd.Process != nil {
+			cmd.Process.Kill()
+		}
+	}, nil
+}
+
+func writeTempFile(pattern string, content string) (*os.File, error) {
+	f, err := ioutil.TempFile("", pattern)
+	if err != nil {
+		return nil, err
+	}
+	defer f.Close()
+
+	if _, err := f.WriteString(content); err != nil {
+		return nil, err
+	}
+
+	return f, nil
+}
+
+func rawConnect(ctx context.Context, t *testing.T, host string, port string) {
+	for {
+		select {
+		case <-ctx.Done():
+			t.Fatalf("failed connecting to %s:%s", host, port)
+		default:
+			conn, err := net.DialTimeout("tcp", net.JoinHostPort(host, port), time.Second)
+			if err != nil {
+				continue
+			}
+			if conn != nil {
+				conn.Close()
+				return
+			}
+		}
+	}
+
+}
